Updated Data layer

+++ app/src/main/java/com/revolt/test/currency_converter/data/entity/CurrencyEntity.kt
@@ -12,5 +12,5 @@ data class CurrencyEntity(
 
     @SerializedName("rates")
     @Expose
-    val rates: HashMap<String, Int>?
+    val rates: HashMap<String, Double>?
 )
